+++ Wire/ReflectionEx.cs
@@ -1,5 +1,7 @@
 ﻿using System;
+using System.Collections.Concurrent;
 using System.Collections.Generic;
+using System.IO;
 using System.Linq;
 using System.Reflection;
 #if SERIALIZATION
@@ -84,6 +86,63 @@ namespace Wire
         public static bool IsOneDimensionalPrimitiveArray(this Type type)
         {
             return type.IsArray && type.GetArrayRank() == 1 && type.GetElementType().IsWirePrimitive();
+        }
+
+        private static readonly ConcurrentDictionary<byte[], Type> TypeNameLookup =
+            new ConcurrentDictionary<byte[], Type>(new ByteArrayEqualityComparer());
+
+        public static byte[] GetTypeManifest(IReadOnlyCollection<byte[]> fieldNames)
+        {
+            IEnumerable<byte> result = new[] { (byte)fieldNames.Count };
+            foreach (var name in fieldNames)
+            {
+                var encodedLength = BitConverter.GetBytes(name.Length);
+                result = result.Concat(encodedLength);
+                result = result.Concat(name);
+            }
+            var versionTolerantHeader = result.ToArray();
+            return versionTolerantHeader;
+        }
+
+        private static Type GetTypeFromManifestName(Stream stream, DeserializerSession session)
+        {
+            var bytes = stream.ReadLengthEncodedByteArray(session);
+
+            return TypeNameLookup.GetOrAdd(bytes, b =>
+            {
+                var shortName = StringEx.FromUtf8Bytes(b, 0, b.Length);
+                var typename = Utils.ToQualifiedAssemblyName(shortName);
+                return Type.GetType(typename, true);
+            });
+        }
+
+        public static Type GetTypeFromManifestFull(Stream stream, DeserializerSession session)
+        {
+            var type = GetTypeFromManifestName(stream, session);
+            session.TrackDeserializedType(type);
+            return type;
+        }
+
+        public static Type GetTypeFromManifestVersion(Stream stream, DeserializerSession session)
+        {
+            var type = GetTypeFromManifestName(stream, session);
+
+            var fieldCount = stream.ReadByte();
+            for (var i = 0; i < fieldCount; i++)
+            {
+                var fieldName = stream.ReadLengthEncodedByteArray(session);
+
+            }
+
+            session.TrackDeserializedTypeWithVersion(type, null);
+            return type;
+        }
+
+        public static Type GetTypeFromManifestIndex(Stream stream, DeserializerSession session)
+        {
+            var typeId = stream.ReadUInt16(session);
+            var type = session.GetTypeFromTypeId(typeId);
+            return type;
         }
     }
