+++ Wire/ValueSerializers/ObjectSerializer.cs
@@ -1,9 +1,7 @@
 using System;
-using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Threading;
-using ByteArrayTypeLookup = System.Collections.Concurrent.ConcurrentDictionary<byte[], System.Type>;
 
 namespace Wire.ValueSerializers
 {
@@ -13,28 +11,12 @@ namespace Wire.ValueSerializers
         public const byte ManifestFull = 255;
         public const byte ManifestIndex = 254;
 
-        private static readonly ByteArrayTypeLookup TypeNameLookup =
-            new ByteArrayTypeLookup(new ByteArrayEqualityComparer());
-
         private readonly byte[] _manifest;
         private readonly byte[] _manifestWithVersionInfo;
 
         private volatile bool _isInitialized;
         private ObjectReader _reader;
         private ObjectWriter _writer;
-
-        private byte[] GetTypeManifest(IReadOnlyCollection<byte[]> fieldNames)
-        {
-            IEnumerable<byte> result = new[] { (byte)fieldNames.Count };
-            foreach (var name in fieldNames)
-            {
-                var encodedLength = BitConverter.GetBytes(name.Length);
-                result = result.Concat(encodedLength);
-                result = result.Concat(name);
-            }
-            var versionTolerantHeader = result.ToArray();
-            return versionTolerantHeader;
-        }
 
         public ObjectSerializer(Type type)
         {
@@ -46,11 +28,11 @@ namespace Wire.ValueSerializers
             var typeName = type.GetShortAssemblyQualifiedName();
             // ReSharper disable once PossibleNullReferenceException
             // ReSharper disable once AssignNullToNotNullAttribute
-            var typeNameBytes = Utils.StringToBytes(typeName);
+            var typeNameBytes = typeName.ToUtf8Bytes();
 
             var fields = ReflectionEx.GetFieldInfosForType(type);
-            var fieldNames = fields.Select(field => Utils.StringToBytes(field.Name)).ToList();
-            var versionInfo = GetTypeManifest(fieldNames);
+            var fieldNames = fields.Select(field => field.Name.ToUtf8Bytes()).ToList();
+            var versionInfo = TypeEx.GetTypeManifest(fieldNames);
 
             //precalculate the entire manifest for this serializer
             //this helps us to minimize calls to Stream.Write/WriteByte 
@@ -122,47 +104,6 @@ namespace Wire.ValueSerializers
             _reader = reader;
             _writer = writer;
             _isInitialized = true;
-        }
-
-        private static Type GetTypeFromManifestName(Stream stream, DeserializerSession session)
-        {
-            var bytes = stream.ReadLengthEncodedByteArray(session);
-
-            return TypeNameLookup.GetOrAdd(bytes, b =>
-            {
-                var shortName = Utils.BytesToString(b, 0, b.Length);
-                var typename = Utils.ToQualifiedAssemblyName(shortName);
-                return Type.GetType(typename, true);
-            });
-        }
-
-        public static Type GetTypeFromManifestFull(Stream stream, DeserializerSession session)
-        {
-            var type = GetTypeFromManifestName(stream, session);
-            session.TrackDeserializedType(type);
-            return type;
-        }
-
-        public static Type GetTypeFromManifestVersion(Stream stream, DeserializerSession session)
-        {
-            var type = GetTypeFromManifestName(stream, session);
-
-            var fieldCount = stream.ReadByte();
-            for (int i = 0; i < fieldCount; i++)
-            {
-                var fieldName = stream.ReadLengthEncodedByteArray(session);
-
-            }
-
-            session.TrackDeserializedTypeWithVersion(type, null);
-            return type;
-        }
-
-        public static Type GetTypeFromManifestIndex(Stream stream, DeserializerSession session)
-        {
-            var typeId = stream.ReadUInt16(session);
-            var type = session.GetTypeFromTypeId(typeId);
-            return type;
         }
     }
 }
