+++ Wire/Wire.csproj
@@ -63,6 +63,7 @@
     <Compile Include="DeserializeSession.cs" />
     <Compile Include="SerializerOptions.cs" />
     <Compile Include="StreamExtensions.cs" />
+    <Compile Include="StringEx.cs" />
     <Compile Include="Surrogate.cs" />
     <Compile Include="TrackingExtensions.cs" />
     <Compile Include="Utils.cs" />
